HW4; task 1; parts 1, 2, 3 made.

+++ src/pro/sky/java/course2/hw4/Train.java
@@ -12,7 +12,7 @@ public class Train extends Transport{
                  int productionYear,
                  String productionCountry,
                  String carcassColor,
-                 short maxSpeed,
+                 int maxSpeed,
                  double tripCost,
                  double tripTime,
                  String departureStation,
